requirements.txt
alembic==1.16.5
fastapi==0.135.1
httpx==0.28.1
itsdangerous==2.2.0
jinja2==3.1.6
openai==1.109.1
pandas==3.0.1
openpyxl==3.1.5
python-docx==1.2.0
psycopg[binary]==3.3.3
python-multipart==0.0.22
sqlalchemy==2.0.48
uvicorn==0.42.0
pytest==8.4.2
